refactor keto use case and use generic function
